Move state machines into their relevant pacakges
