+++ src/bwyap/familyfeud/game/FamilyFeudGame.java
@@ -3,6 +3,7 @@ package bwyap.familyfeud.game;
 import java.util.List;
 
 import bwyap.familyfeud.game.state.FFState;
+import bwyap.familyfeud.game.state.FFStateMachine;
 import bwyap.familyfeud.game.state.FFStateType;
 import bwyap.familyfeud.testdriver.FamilyFeudTestDriver;
 import bwyap.utility.logging.Logger;
