+++ src/bwyap/familyfeud/game/state/StatePlay.java
@@ -1,9 +1,9 @@
 package bwyap.familyfeud.game.state;
 
-import bwyap.familyfeud.game.FFPlayStateMachine;
 import bwyap.familyfeud.game.FamilyFeudGame;
 import bwyap.familyfeud.game.InvalidDataException;
 import bwyap.familyfeud.game.play.state.FFPlayState;
+import bwyap.familyfeud.game.play.state.FFPlayStateMachine;
 import bwyap.familyfeud.game.play.state.FFPlayStateType;
 
 /**
